+++ Graphics/Blank/DeviceContext.hs
@@ -21,7 +21,7 @@ data DeviceContext = DeviceContext
         , eventQueue  :: EventQueue                 -- ^ A single (typed) event queue
         , ctx_width   :: !Int
         , ctx_height  :: !Int
-        , ctx_devicePixelRatio :: !Int
+        , ctx_devicePixelRatio :: !Float
         }
 
 instance Image DeviceContext where 
@@ -37,7 +37,7 @@ deviceCanvasContext cxt = CanvasContext 0 (ctx_width cxt) (ctx_height cxt)
 --   to use the native devicePixelRatio, and if successful, 'devicePixelRatio' will return a number other than 1.
 --   You can think of devicePixelRatio as the line width to use to make lines look one pixel wide.
 
-devicePixelRatio ::  DeviceContext -> Int
+devicePixelRatio ::  DeviceContext -> Float
 devicePixelRatio = ctx_devicePixelRatio
 
 -- | internal command to send a message to the canvas.
